Refactor: 채팅 페이지 헤더 컴포넌트 분리 (#48)
