+++ src/features/chatting/containers/ChattingContainer.tsx
@@ -1,8 +1,6 @@
 import { type ReactNode, type Dispatch, type SetStateAction, Fragment, useRef } from "react";
 import { Platform, ScrollView } from "react-native";
-import { useRouter } from "expo-router";
 import { Image as Img } from "expo-image";
-import MaterialIcons from "@expo/vector-icons/MaterialIcons";
 import styled from "styled-components/native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import Loading from "@/src/components/ui/Loading";
@@ -15,6 +13,7 @@ import AiChatBox from "../components/AiChatBox";
 import UserChatBox from "../components/UserChatBox";
 import { useChatting } from "../hooks/useChatting";
 import type { TThreadDate } from "../types/chattingTypes";
+import ChatHeader from "../components/ChatHeader";
 
 type ChattingContainerProps = {
   threadDate: TThreadDate;
@@ -33,7 +32,6 @@ export default function ChattingContainer({
   renderAssistantList,
   readonly,
 }: ChattingContainerProps) {
-  const router = useRouter();
   const scrollViewRef = useRef<ScrollView>(null);
 
   const {
@@ -62,15 +60,12 @@ export default function ChattingContainer({
     <Fragment>
       {renderAssistantList({ isVisible, setIsVisible, onPressAiCard: handlePressAiCard })}
       <SafeView edges={["left", "right", "top"]}>
-        <HeaderBox>
-          <BackButton onPress={() => router.back()} hitSlop={7}>
-            <MaterialIcons name="arrow-back-ios" size={28} color={theme.colors.black} />
-          </BackButton>
-          <HeaderButton onPress={handleHeaderPress} hitSlop={7} disabled={readonly}>
-            <Image src={data.result.aiProfileImageS3} />
-            <AiNameText>{data.result.aiProfileName}</AiNameText>
-          </HeaderButton>
-        </HeaderBox>
+        <ChatHeader
+          imageSrc={data.result.aiProfileImageS3}
+          assistantName={data.result.aiProfileName}
+          onPress={handleHeaderPress}
+          readonly={readonly}
+        />
         <ScrollBox
           ref={scrollViewRef}
           onContentSizeChange={() => {
@@ -110,41 +105,6 @@ export default function ChattingContainer({
 const SafeView = styled(SafeAreaView)`
   flex: 1;
   background-color: ${({ theme }) => theme.colors.white};
-`;
-
-const HeaderBox = styled.View`
-  width: 100%;
-  height: ${responsiveToPxByHeight("110px")};
-  background-color: ${({ theme }) => theme.colors.white};
-  flex-direction: row;
-  padding: 0px ${responsiveToPx("24px")};
-  align-items: center;
-  gap: ${({ theme }) => theme.gap.lg};
-`;
-
-const BackButton = styled.TouchableOpacity`
-  width: ${responsiveToPx("28px")};
-  height: ${responsiveToPx("28px")};
-  justify-content: center;
-  align-items: center;
-`;
-
-const HeaderButton = styled.TouchableOpacity`
-  flex-direction: row;
-  gap: ${({ theme }) => theme.gap.lg};
-  align-items: center;
-`;
-
-const Image = styled(CachedImage)`
-  width: ${responsiveToPx("48px")};
-  height: ${responsiveToPx("48px")};
-  border-radius: 9999px;
-`;
-
-const AiNameText = styled.Text`
-  color: ${({ theme }) => theme.colors.black};
-  font-family: ${({ theme }) => theme.fontFamily.nsBold};
-  font-size: ${({ theme }) => theme.fontSize.lg};
 `;
 
 const ScrollBox = styled(ScrollView)`
